Break site generation into functions to allow subtemplates.
Main site generation is now contained in the generate_site() function,
which can respond to certain tokens by calling content generation
functions. These helper functions may run progressive token
replacements, as with replace_tokens(), or call on a subtemplate to
generate a sequence of lines, as with build_nav(). Notice how
build_nav() creates a list of lines, and then collapses it to a string
using "".join().

Since the helper functions take the page data as an argument, they can
change their own behavior based on the JSON data. build_nav(), for
instance, loads a nav_template if one is specified, but defaults to
"input_dir/nav_template.html" if not.